moved student tests out of the integration/full test
