+++ cypress/disabled/student_tests/group_chooser_spec.js
