+++ cypress/disabled/student_tests/image_tool_spec.js
