+++ cypress/disabled/student_tests/left_nav_test_spec.js
