+++ cypress/disabled/student_tests/text_tool_spec.js
